FDOS-509 FDOS 509 renamed the folders and files
