+++ tests/service_automation/tests/features/crud_api_features/location_api.feature
@@ -1,4 +1,4 @@
-@fdos-location-api
+@crud-location-api
 Feature: API tests for Fdos Location resources
