+++ tests/service_automation/tests/step_definitions/crud_api_steps/test_healthcare_service_api.py
@@ -1,5 +1,4 @@
 """
-from http import HTTPStatus
 
 from pytest_bdd import given, parsers, scenarios, then, when
 from step_definitions.common_steps.data_steps import *  # noqa: F403
@@ -9,14 +8,7 @@ from utilities.infra.dns_util import wait_for_dns
 
 
 # Load feature file
-scenarios("./fdos_crud_features/location_api.feature")
-
-
-@given(parsers.re(r'the dns for "(?P<api_name>.*?)" is resolvable'))
-def dns_resolvable(api_name, env, workspace):
-    r53 = get_r53(workspace, api_name, env)
-    assert wait_for_dns(r53)
-
+scenarios("./crud_api_features/healthcare_service_api.feature")
 
 @when(
     parsers.re(r'I request data from the "(?P<api_name>.*?)" endpoint "(?P<resource_name>.*?)"'),
@@ -25,9 +17,8 @@ def dns_resolvable(api_name, env, workspace):
 def send_get(api_request_context_mtls_crud, api_name, resource_name):
     url = get_url(api_name) + "/" + resource_name
     # Handle None or empty params
-    #  response = api_request_context_mtls_crud.get(url)
-    return HTTPStatus.OK
-
+    response = api_request_context_mtls_crud.get(url)
+    return response
 
 @then(parsers.parse('I receive a status code "{status_code:d}" in response'))
 def status_code(fresponse, status_code):
